Store hashed password in cookie to avoid re-hashing at every request
The only drawback is that if admins change a user's password from
plaintext to hashed or vice versa that user will have to log in again.
Attackers knowing hashed passwords shouldn't be a problem as the server
only accepts it if it's signed (together with a recent timestamp) with
its secret key.

Fixes #1076.