Reload the admin/contest page only once once when the contest ends
Previously the callback to update remaining time in both CWS and AWS
would continue to fire every second even after the page reload was
triggered. If the server was slow to respond, the reload would be issued
again, resulting in an infinite loop and possibly DoS-ing the server.

Fix this by cancelling the update timer when a page reload is issued.

Fixes #1104.